+++ Cis.Model/StockCard.cs
@@ -7,8 +7,8 @@ namespace Cis.Model
         public string ProductName { get; set; }
         public string BatchId { get; set; }
         public string BatchCode { get; set; }
-        public decimal QtyOnHand { get; set; }
-        public decimal QtyReserved { get; set; }
+        public int QtyOnHand { get; set; }
+        public int QtyReserved { get; set; }
 
         public Product Product { get; set; }
     }
